+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [Unreleased]
+* Add basic support for matching on extensible types.
 * Add the attribute `@coq_type_annotation` to generate the type annotation of an expression.
 * Show nicer error messages for errors in the configuration file.
 * Remove existential types from the modules (except for the first-class modules).